chore: put export-type enum in separate file
AB#31009

To be able to use ExportType enum easily in Portalicious, we need to separate it into a file of its own. This should be in line with the 121-service guidelines so it shouldn't be an issue, and is worth the refactor on its own IMHO.
