+++ services/121-service/src/actions/action.entity.ts
@@ -1,7 +1,7 @@
 import { Column, Entity, ManyToOne, Relation } from 'typeorm';
 
 import { Base121AuditedEntity } from '@121-service/src/base-audited.entity';
-import { ExportType } from '@121-service/src/metrics/dto/export-details.dto';
+import { ExportType } from '@121-service/src/metrics/enum/export-type.enum';
 import { ProgramEntity } from '@121-service/src/programs/program.entity';
 import { UserEntity } from '@121-service/src/user/user.entity';
