+++ services/121-service/src/metrics/enum/export-type.enum.ts
@@ -1,0 +1,10 @@
+export enum ExportType {
+  allPeopleAffected = 'all-people-affected',
+  included = 'included',
+  payment = 'payment',
+  unusedVouchers = 'unused-vouchers',
+  vouchersWithBalance = 'vouchers-with-balance',
+  toCancelVouchers = 'to-cancel-vouchers',
+  duplicates = 'duplicates',
+  cardBalances = 'card-balances',
+}
